Rename categories to citar-* and name maps accordingly
This is a matter of style. If the categories are citar-specific, it is better to
use package prefixes. In Consult I also use consult-* categories. In contrast,
in Marginalia we use categories without prefixes since the categories defined
there are supposed to be more general and extent to existing Emacs entities.